fix[doc]: docstring for usage of ShotProps.spin_drift
The docstring on `ShotProps.spin_drift` in `shot.py` says it
returns the spin drift in inches, but that's not correct.
Internally, it divides by 12 to get feet, because that's the
units used by the `adjusted_range: Vector`.

The documentation for `BCLIBC_ShotProps::spin_drift` in
`base_types.cpp` correctly notes the return type being in feet,
so no changes were needed there.