Fix cpp (#296)
* fix: avoid cpp warnings
* fix: includes fix

* refactor: better WindSock initialisation and Wind conversion from PyObject to C++ object

* fix[doc]: docstring for usage of ShotProps.spin_drift (#299)

The docstring on `ShotProps.spin_drift` in `shot.py` says it
returns the spin drift in inches, but that's not correct.
Internally, it divides by 12 to get feet, because that's the
units used by the `adjusted_range: Vector`.

The documentation for `BCLIBC_ShotProps::spin_drift` in
`base_types.cpp` correctly notes the return type being in feet,
so no changes were needed there.

* fix: remove trailing newline

---------

Co-authored-by: Billy McCulloch <billy.mcculloch@lynntech.com>

* refactor: avoid multiple rethrows when handling c++ exceptions to Cython (#300)
* fix: BCLIBC_WindSock move constructor
* refactor: uses operator[]
* chore: update CHANGELOG.md

---------

Co-authored-by: Billy McCulloch <bjmcculloch@gmail.com>
Co-authored-by: Billy McCulloch <billy.mcculloch@lynntech.com>